+++ CHANGELOG.md
@@ -1,3 +1,32 @@
+# [2.43.0](https://github.com/coinranking/exchanges/compare/v2.42.1...v2.43.0) (2020-05-27)
+
+
+### Features
+
+* excoincial driver ([4b2e508](https://github.com/coinranking/exchanges/commit/4b2e50811fe885bd9d59c047c0eec5b388b39be7))
+
+## [2.42.1](https://github.com/coinranking/exchanges/compare/v2.42.0...v2.42.1) (2020-05-22)
+
+
+### Bug Fixes
+
+* fix the base and quote volumes of ABIT ([c379097](https://github.com/coinranking/exchanges/commit/c37909767d390a0a21c90850c5245684b81235d7))
+* fix the bcbitcoin test, because it failed after 24 hours ([52884fa](https://github.com/coinranking/exchanges/commit/52884fa9c97f183c5395676c616a8926c271f5e1))
+
+# [2.42.0](https://github.com/coinranking/exchanges/compare/v2.41.0...v2.42.0) (2020-05-22)
+
+
+### Features
+
+* update abit with new url ([c04fe68](https://github.com/coinranking/exchanges/commit/c04fe684644573d7b143773d5034dae20215e2d8))
+
+# [2.41.0](https://github.com/coinranking/exchanges/compare/v2.40.0...v2.41.0) (2020-05-21)
+
+
+### Features
+
+* add support for BC Bitcoin ([db140ed](https://github.com/coinranking/exchanges/commit/db140edab2b9afd0630ffaf08eb9057042d05222))
+
 # [2.40.0](https://github.com/coinranking/exchanges/compare/v2.39.0...v2.40.0) (2020-05-14)